% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AW_Novelty.R
\name{AW_calc}
\alias{AW_calc}
\title{Compute an abundance weighted FPD}
\usage{
AW_calc(focal.species, community.data, fpd.mat, metric = c("NND", "MPD"),
  na.rm = TRUE)
}
\arguments{
\item{focal.species}{The focal species or plot identifier that
you want to use to create a the phylogeny}

\item{community.data}{The community data file. Each observation
of a species in a community should be its own row. For now,
the column specifiy which community is which is called 
\code{exotic_species}, but this will be generalized later
using an argument to specifiy which column marks community 
separators and which column holds the actual species names}

\item{fpd.mat}{A functional-phylogenetic distance matrix}

\item{metric}{The metric of novelty you'd like to calculate.
Currently supports mean pairwise distance and nearest neighbor
distance}

\item{na.rm}{A logical indicating whether to remove missing
values}
}
\value{
A numeric
}
\description{
This function is designed to be called internally
by \code{rarefy_FPD} and weights each species distance based on
their local abundance.
}
\author{
Sam Levin
}
