% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_phylo_dist.R
\name{make_local_phylo_dist}
\alias{make_local_phylo_dist}
\title{Make Local Scale Phylogenetic Distance Matrix}
\usage{
make_local_phylo_dist(focal.species, community.data, phylo)
}
\arguments{
\item{focal.species}{The focal species or plot identifier that
you want to use to create a the phylogeny}

\item{community.data}{The community data file. Each observation
of a species in a community should be its own row. For now,
the column specifiy which community is which is called 
\code{exotic_species}, but this will be generalized later
using an argument to specifiy which column marks community 
separators and which column holds the actual species names}

\item{phylo}{A larger phylogeny that contains all of the 
species in the \code{community.data} file.}
}
\value{
A phylogenetic distance matrix in the form of a
\code{data.frame}
}
\description{
Function for creating a phylogenetic distance matrix
from data that does not require you to subset first.
}
\author{
Sam Levin
}
