% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_fun_dist.R
\name{make_local_trait_dist}
\alias{make_local_trait_dist}
\title{Make a functional trait distance matrix}
\usage{
make_local_trait_dist(focal.species, community.data, trait.data, traits, scale)
}
\arguments{
\item{focal.species}{The focal species or plot identifier that
you want to use to create a the phylogeny}

\item{community.data}{The community data file. Each observation
of a species in a community should be its own row. For now,
the column specifiy which community is which is called 
\code{exotic_species}, but this will be generalized later
using an argument to specifiy which column marks community 
separators and which column holds the actual species names}

\item{trait.data}{A data frame that contains trait data information.
Currently, this only supports names from \emph{Levin et al 2017}, but
will hopefully soon be generalized to include all trait names listed
in \emph{TRY (Kattge et al 2017)}.}

\item{traits}{A character vector of trait names that should match the names
in \code{trait.data}. An effect method of subsetting this is to simply call
\code{names(trait.data)[first.trait:last.trait]}}

\item{scale}{The spatial scale at which to analyze the data}
}
\value{
A distance matrix of class \code{dist}.
}
\description{
Creates a Gower distance matrix using the method
of Pavoine et al 2009 to handle many types of data
}
\note{
These
can easily be coerced to other data structures, but must
be coerced to a \code{matrix} first.
}
\author{
Sam Levin
}
