% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regional_phylo_dist.R
\name{make_regional_phylo_dist}
\alias{make_regional_phylo_dist}
\title{Make a species pool-level phylogenetic distance matrix}
\usage{
make_regional_phylo_dist(spp.vector, phylo, square_root = TRUE)
}
\arguments{
\item{spp.vector}{A character vector of species names. Species names
must match those in phylogeny.}

\item{phylo}{A larger phylogeny that contains all of the 
species in the \code{spp.vector}. If this condition is not met,
the function will try to merge in the missing species using
\code{pez::congeneric.merge} while warning the user.}

\item{square_root}{Square root transform branch lengths in distance matrix?}
}
\value{
A phylogenetic distance matrix in the form of a
\code{data.frame}
}
\description{
Create a phylogeny from a species pool and then calculate
patristic distances between all species
}
\author{
Sam Levin
}
