% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regional_fun_dist.R
\name{make_regional_trait_dist}
\alias{make_regional_trait_dist}
\title{Make a regional scale functional distance matrix}
\usage{
make_regional_trait_dist(trait.data, traits)
}
\arguments{
\item{trait.data}{A data frame that contains trait data information.
Currently, this only supports names from \emph{Levin et al 2017}, but
will hopefully soon be generalized to include all trait names listed
in \emph{TRY (Kattge et al 2017)}.}

\item{traits}{A character vector of trait names that should match the names
in \code{trait.data}. An effect method of subsetting this is to simply call
\code{names(trait.data)[first.trait:last.trait]}}
}
\value{
An object of class \code{dist}
}
\description{
This function is very similar to 
\code{make_local_traits_distance}, but will not subset to plot/focal species
level by default.
}
\note{
These
can easily be coerced to other data structures, but must
be coerced to a \code{matrix} first.
}
\seealso{
\code{\link{make_local_trait_dist}}, \code{\link{make_regional_trait_ktab}}
}
\author{
Sam Levin
}
