% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regional_KTab.R
\name{make_regional_traits_ktab}
\alias{make_regional_traits_ktab}
\title{Make a KTab for a whole data set}
\usage{
make_regional_traits_ktab(trait.data, traits)
}
\arguments{
\item{trait.data}{A data frame that contains trait data information.
Currently, this only supports names from \emph{Levin et al 2017}, but
will hopefully soon be generalized to include all trait names listed
in \emph{TRY (Kattge et al 2017)}.}

\item{traits}{A character vector of trait names that should match the names
in \code{trait.data}. An effect method of subsetting this is to simply call
\code{names(trait.data)[first.trait:last.trait]}}
}
\value{
A list consisting of 
\itemize{
   \item{\code{KTab}}{The \code{ktab} which can be passed to 
   \code{dist.ktab}}
   \item{\code{VarTypes}}{A character vector that can be passed to 
   \code{dist.ktab}}
}
}
\description{
This is very similar to 
\code{make_local_traits_ktab}, but does not subset the data
by default.
}
\author{
Sam Levin
}
