% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_R2.R
\name{r2_calc}
\alias{r2_calc}
\title{Extract R^2 from a linear model}
\usage{
r2_calc(data, formula)
}
\arguments{
\item{data}{The data frame to perform the calculation on}

\item{formula}{A formula for the model}
}
\value{
A scalar numeric representing the adjusted r-squared for the model
}
\description{
Calculate the adjusted r-squared from a linear model given
a dataframe and formula
}
\author{
Sam Levin
}
