% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rarefy_FPD.R
\name{rarefy_FPD}
\alias{rarefy_FPD}
\title{Rarefied Function-Phylogenetic Distances}
\usage{
rarefy_FPD(focal.species, phylo.mat, fun.mat, metric = c("MPD", "NND"),
  n.resamp = 1000, n.rare, a, p, abundance.weighted = FALSE,
  community.data = NULL, log = TRUE)
}
\arguments{
\item{focal.species}{The name of the focal species in the 
community to be rarefied.}

\item{phylo.mat, fun.mat}{A phylogenetic or functional distance
matrix with species names as row names and column names. 
Can either be a \code{data frame}, \code{matrix}, or \code{dist}.}

\item{metric}{Either MPD (\code{MPD}) or nearest neighbor distance
(\code{NND}). The default is to calculate both metrics.}

\item{n.resamp}{The number of rarefied resamplings to perform. 
Default is 1000.}

\item{n.rare}{The number of species to rarefy the community down to.
Must be less than the total in the community.}

\item{a}{The phylogenetic scaling factor for the calculation of
functional-phylogenetic distances}

\item{p}{The power to raise each distance value to in the functional-
phylogenetic distance calculation}

\item{abundance.weighted}{A logical value indicating whether or not to 
weight species by their relative abundances}

\item{community.data}{The community data file. Each observation
of a species in a community should be its own row. For now,
the column specifiy which community is which is called 
\code{exotic_species}, but this will be generalized later
using an argument to specifiy which column marks community 
separators and which column holds the actual species names}

\item{log}{log transform the output?}
}
\value{
A list with 4 components
\itemize{
\item{\code{rare.mpd}}{Rarefied MPD value for the species in the community}
\item{\code{sample.mpd}}{The values from each iteration of the sampling}
\item{\code{rare.nnd}}{Rarefied NND values for the species in the community}
\item{\code{sample.nnd}}{The values from each iteration of the sampling}#' 
}
}
\description{
\code{rare_FPD} resamples a phylogenetic and functional distance matrix
a set number of times so that communities of different sizes can be 
compared to each other.
}
\author{
Sam Levin
}
