% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AW_Min.R
\name{weighted.min}
\alias{weighted.min}
\title{Calculate abundance weighted NND}
\usage{
weighted.min(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{a vector or matrix of numerical values}

\item{w}{a vector or matrix of numerical weights whose positions
correspond to the values in \code{x}.}

\item{na.rm}{A logical indicating whether to ignore missing values.}
}
\value{
A numerical value
}
\description{
Calculates abundance weighted minimum
}
\seealso{
\code{\link[stats]{weighted.mean}}
}
\author{
Sam Levin
}
