% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rarefy_FPD.R
\name{rarefy_FPD}
\alias{rarefy_FPD}
\title{Rarefied Function-Phylogenetic Distances}
\usage{
rarefy_FPD(focal.species, phylo.mat, fun.mat, metric = c("MPD", "NND"),
  n.resamp = 1000, n.rare, a, p)
}
\arguments{
\item{focal.species}{The name of the focal species in the 
community to be rarefied.}

\item{phylo.mat, fun.mat}{A phylogenetic or functional distance
matrix with species names as row names and column names. 
Can either be a \code{data frame}, \code{matrix}, or \code{dist}.}

\item{metric}{Either MPD (\code{MPD}) or nearest neighbor distance
(\code{NND}). The default is to calculate both metrics.}

\item{n.resamp}{The number of rarefied resamplings to perform. 
Default is 1000.}

\item{n.rare}{The number of species to rarefy the community down to.
Must be less than the total in the community.}

\item{a}{The phylogenetic scaling factor for the calculation of
functional-phylogenetic distances}

\item{p}{The power to raise each distance value to in the functional-
phylogenetic distance calculation}
}
\description{
\code{rare_FPD} resamples a phylogenetic and functional distance matrix
a set number of times so that communities of different sizes can be 
compared to each other.
}
\author{
Sam Levin
}
